Updates/Flexibility
Note: Last commit should have stated that the logo was resized to reduce
the size of the header.  Nav in header and footer were made a little
bigger and smaller to fit their positioning.

This commit refers to every page utilizing percentages to position items
from the left margin in order to make the pages more responsive during
window resizing.  The only page not changed in this commit is the
projects showcase page.